Port test suite from mocha to vitest
Replace mocha with vitest for test running. Convert the test file to ESM (.mjs) since vitest only supports ESM imports, swap assert.strictEqual for expect().toBe(), and replace done() callbacks with returned Promises. Move the per-test timeout from this.timeout() to vitest's third it() argument.

Update the prettier glob to also cover .mjs files.
