+++ package.json
@@ -16,9 +16,9 @@
   "scripts": {
     "mksnapshot": "node ./mksnapshot.js",
     "install": "node ./download-mksnapshot.js",
-    "prettier": "prettier --single-quote --no-semi --check \"**/*.js\"",
-    "prettier:write": "prettier --single-quote --no-semi --write \"**/*.js\"",
-    "test": "mocha && npm run prettier"
+    "prettier": "prettier --single-quote --no-semi --check \"**/*.{js,mjs}\"",
+    "prettier:write": "prettier --single-quote --no-semi --write \"**/*.{js,mjs}\"",
+    "test": "vitest run && npm run prettier"
   },
   "license": "MIT",
   "dependencies": {
@@ -29,9 +29,9 @@
   },
   "devDependencies": {
     "@electron/fiddle-core": "^1.3.0",
-    "mocha": "^10.1.0",
     "prettier": "^3.6.2",
-    "semver": "^7.3.8"
+    "semver": "^7.3.8",
+    "vitest": "^2.1.8"
   },
   "standard": {
     "ignore": [
